+++ src/irc.c
@@ -156,6 +156,7 @@ static struct numeric_events_tag {
 	{ "317", "RPL_WHOISIDLE",           NO_WINDOW,      0, event_whois_idle },
 	{ "318", "RPL_ENDOFWHOIS",          ACTIVE_WINDOW,  2, NULL },
 	{ "319", "RPL_WHOISCHANNELS",       NO_WINDOW,      0, event_whois_channels },
+	{ "320", "",                        NO_WINDOW,      0, event_whois_ssl },
 	{ "321", "RPL_LISTSTART",           NO_WINDOW,      0, event_liststart },
 	{ "322", "RPL_LIST",                NO_WINDOW,      0, event_list },
 	{ "323", "RPL_LISTEND",             STATUS_WINDOW,  1, NULL },